Adding serialization package, restricting Eigen serializer.
